+++ src/lib/logging/core/types.ts
@@ -1,5 +1,3 @@
-import type { Formatter } from "./formatter.js";
-
 export const LogLevel = {
   debug: 10,
   info: 20,
@@ -12,8 +10,6 @@ export type LogLevelType = keyof typeof LogLevel;
 
 export type LogMessageType = string | object | number | boolean;
 
-export type DateFmtFnType = () => string;
-
 export type LogDataType = {
   prefix: string;
   level: LogLevelType;
@@ -22,24 +18,4 @@ export type LogDataType = {
   column?: string;
   fileName?: string;
   method?: string;
-};
-
-export type ProviderOptions = {
-  level?: LogLevelType;
-  formatter?: Formatter;
-};
-
-export interface SizeBasedPolicyType {
-  type: "size";
-  maxSize?: number;
-}
-
-export interface TimeBasedPolicyType {
-  type: "time";
-  instant: "hour" | "day" | "week" | "month";
-  duration: number;
-}
-
-export type FileProviderOptions = ProviderOptions & {
-  policy?: SizeBasedPolicyType | TimeBasedPolicyType;
 };
