+++ src/lib/logging/formatter/types.ts
@@ -1,0 +1,1 @@
+export type DateFmtFnType = () => string;
