+++ src/lib/logging/index.ts
@@ -1,3 +1,5 @@
+export { AbstractLogger, Logger } from "./core/index.js";
+
 export {
   BaseFormatter,
   dmyFormat,
@@ -6,23 +8,26 @@ export {
   isoDateTimeFormat,
   JSONFormatter,
   mdyFormat,
-} from "./core/formatter.js";
-export {
-  AbstractLogger,
-  ConsoleProvider,
-  FileProvider,
-  Logger,
-  LoggerProvider,
-} from "./core/index.js";
+} from "./formatter/index.js";
 
 export {
-  DateFmtFnType,
-  FileProviderOptions,
+  ConsoleProvider,
+  FileProvider,
+  LoggerProvider,
+} from "./provider/index.js";
+
+export {
   LogDataType,
   LogLevel,
   LogLevelType,
   LogMessageType,
+} from "./core/types.js";
+
+export {
+  FileProviderOptions,
   ProviderOptions,
   SizeBasedPolicyType,
   TimeBasedPolicyType,
-} from "./core/types.js";
+} from "./provider/types.js";
+
+export { DateFmtFnType } from "./formatter/types.js";
